+++ src/FlexiChains.jl
@@ -38,6 +38,7 @@ include("interface/mergesubset.jl")
 include("interface/samples.jl")
 include("interface/rand.jl")
 include("varname.jl")
+include("flatten.jl")
 include("conversions.jl")
 include("stancsv.jl")
 
@@ -66,8 +67,8 @@ include("plots.jl")
 @public mrankplot, mrankplot!
 
 # Convenience re-exports.
-using DimensionalData: At, Near, Contains, (..), Touches, Where, Not
-export At, Near, Contains, (..), Touches, Where, Not
+using DimensionalData: At, Near, Contains, (..), Touches, Where, Not, DimArray
+export At, Near, Contains, (..), Touches, Where, Not, DimArray
 using Statistics: mean, median, std, var, quantile
 export mean, median, std, var, quantile
 using StatsBase: summarystats, mad