+++ HISTORY.md
@@ -1,3 +1,11 @@
+# 0.5.4
+
+Added a method, `DimensionalData.DimArray(::FlexiChain, kwargs...)`, to convert a `FlexiChain` into a 3D `DimArray` with dimensions `(iters, chains, parameters)`.
+A method for `Base.Array` has also been added, which is the same but drops the dimensional metadata.
+
+Implemented `PosteriorStats.loo` for chains of log-likelihoods, as well as Turing.jl models.
+Please see the documentation for more information about the PosteriorStats integration.
+
 # 0.5.3
 
 Added the `FlexiChains.rankplot` function for creating rank plots with Plots.jl backend.